+++ R/iso.R
@@ -11,7 +11,7 @@
 #' @param dlim Vector of desired limits of isodistances in metres.
 #' @param heap Type of heap to use in priority queue. Options include
 #' Fibonacci Heap (default; `FHeap`), Binary Heap (`BHeap`),
-#' `Radix`, Trinomial Heap (`TriHeap`), Extended Trinomial Heap
+#' Trinomial Heap (`TriHeap`), Extended Trinomial Heap
 #' (`TriHeapExt`, and 2-3 Heap (`Heap23`).
 #' @return A single `data.frame` of isodistances as points sorted anticlockwise
 #' around each origin (`from`) point, with columns denoting the `from` points